Port convenience, moderation modules to new arch.
